+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Add `zng-ext-setup` crate for creating (un)installers.
+
 * Improve `zng::task::io::Metrics`.
     - `Display` now hides the total estimated length when read/write overshoots it.
     - `Measure::metrics` is now `&self`.